+++ config/config.exs
@@ -1,5 +1,3 @@
 import Config
 
-# Import environment specific config. This must remain at the bottom
-# of this file so it overrides the configuration defined above.
-import_config "#{Mix.env()}.exs"
+if Mix.env() == :test, do: import_config("test.exs")