+++ config/test.exs
@@ -1,3 +1,2 @@
 import Config
-
 config :phoenix, json_library: Jason